split out classes in to a basic package structure
