+++ src/main/java/com/entechconsulting/eventservices/repository/TempEventRepository.java
@@ -1,8 +1,8 @@
-package com.entechconsulting.eventservices;
+package com.entechconsulting.eventservices.repository;
 
 import org.springframework.data.repository.CrudRepository;
 
-import com.entechconsulting.eventservices.TempEvent;
+import com.entechconsulting.eventservices.repository.TempEvent;
 
 public interface TempEventRepository extends CrudRepository<TempEvent,Long>{
 }
